+++ source/Game.cpp
@@ -17,29 +17,7 @@ Game::Game()
 {
 }
 
-// Game specific _______________________________
-void Game::LoadData()
-{
-  // create field
-  for(auto vec : FIELD_POSITION)
-  {
-    new Field(this, vec.x, vec.y);
-  }
-
-  // create path
-  for(auto path : PATH_EDGES)
-  {
-    new Path(this, mFields[path.first], mFields[path.second]);
-  }
-
-  // create players
-  mTBPlayer1 = new TBPlayer(this, mFields[INITIAL_FIELD_FOR_1], true, INITIAL_PENDING_BOMB_NUM, MANUAL_PLAYER);
-  mTBPlayer2 = new RandomPlayer(this, mFields[INITIAL_FIELD_FOR_2], false, INITIAL_PENDING_BOMB_NUM, RANDOM_PLAYER);
-
-  // create TurnManager
-  mTurnManager = new TurnManager(this, mTBPlayer1, mTBPlayer2);
-  //std::cout << this << std::endl;
-}
+// Game::LoadData() is game specific 
 
 void Game::UnloadData()
 {
@@ -53,49 +31,6 @@ void Game::UnloadData()
     SDL_DestroyTexture(i.second);
   }
   mTextures.clear();
-}
-
-void Game::AddField(Field *field)
-{
-  mFields.push_back(field);
-}
-
-void Game::RemoveField(Field *field)
-{
-  auto iter = std::find(mFields.begin(), mFields.end(), field);
-  if(iter != mFields.end())
-  {
-    mFields.erase(iter);
-  }
-}
-
-void Game::AddPath(Path *path)
-{
-  mPaths.push_back(path);
-}
-
-void Game::RemovePath(Path *path)
-{
-  auto iter = std::find(mPaths.begin(), mPaths.end(), path);
-  if(iter != mPaths.end())
-  {
-    mPaths.erase(iter);
-  }
-}
-
-void Game::AddSettedBomb(Bomb *bomb)
-{
-  mSettedBombs.push_back(bomb);
-}
-
-void Game::RemoveSettedBomb(Bomb *bomb)
-{
-  auto iter = std::find(mSettedBombs.begin(), mSettedBombs.end(), bomb);
-  if(iter != mSettedBombs.end())
-  {
-    std::iter_swap(iter, mSettedBombs.end() - 1);
-    mSettedBombs.pop_back();
-  }
 }
 
 // ___________general functions______________
@@ -203,13 +138,14 @@ void Game::UpdateGame()
   
   // Update all actors
   mUpdatingActors = true;
+  for(const auto& actor : mActors) actor->Update(deltaTime);
   // mActors to explicitly chosed actors
-  for(auto actor : mSettedBombs) actor -> Update(deltaTime);
-  mTurnManager->Update(deltaTime);
-  if(mTBPlayer1 != nullptr) mTBPlayer1->Update(deltaTime);
-  if(mTBPlayer2 != nullptr) mTBPlayer2->Update(deltaTime);
-  for(auto field : mFields) field->Update(deltaTime);
-  mUpdatingActors = false;
+  // for(auto actor : mSettedBombs) actor -> Update(deltaTime);
+  // mTurnManager->Update(deltaTime);
+  // if(mTBPlayer1 != nullptr) mTBPlayer1->Update(deltaTime);
+  // if(mTBPlayer2 != nullptr) mTBPlayer2->Update(deltaTime);
+  // for(auto field : mFields) field->Update(deltaTime);
+  // mUpdatingActors = false;
   
   // move penidingActors to Actors
   for(auto pending:mPendingActors) mActors.emplace_back(pending);
