+++ source/Path.cpp
@@ -2,10 +2,11 @@
 #include "Field.h"
 #include "SpriteComponent.h"
 #include "Game.h"
+#include "TimeBombApp.h"
 #include <cmath>
 
-Path::Path(Game *game, Field *nord1, Field *nord2)
-:Actor(game), mNord1(nord1), mNord2(nord2), mWeight(DEFAULT_HEAVY_WEIGHT)
+Path::Path(TimeBombApp *game, Field *nord1, Field *nord2)
+:Actor(game), mApp(game), mNord1(nord1), mNord2(nord2), mWeight(DEFAULT_HEAVY_WEIGHT)
 {
     if(mNord1->GetPosition().y < mNord2->GetPosition().y) std::swap(mNord1, mNord2);
     SpriteComponent *sc = new SpriteComponent(this, 30, MANUAL_POSITIONING, MANUAL_SCALING);
@@ -28,5 +29,5 @@ Path::Path(Game *game, Field *nord1, Field *nord2)
 
 Path::~Path()
 {
-    GetGame()->RemovePath(this);
+    mApp->RemovePath(this);
 }
