+++ source/Players.h
@@ -4,7 +4,7 @@
 class RandomPlayer : public TBPlayer
 {
 public:
-    RandomPlayer(class Game* game, class Field* ip, bool player, const int &bombnum, int playerType = 0);
+    RandomPlayer(class TimeBombApp* game, class Field* ip, bool player, const int &bombnum, int playerType = 0);
     ~RandomPlayer(){};
 
     class Field* ChooseField(std::vector<class Field*> candFields) override;
