+++ source/TBPlayer.h
@@ -11,7 +11,7 @@ enum PlayerType
 class TBPlayer : public Actor
 {
 public:
-    TBPlayer(class Game* game, class Field* ip, bool player, const int &bombnum, int playerType = 0); // ip : initial position
+    TBPlayer(class TimeBombApp* game, class Field* ip, bool player, const int &bombnum, int playerType = 0); // ip : initial position
     ~TBPlayer(); // call the ending manager
 
     void UpdateActor(float deltaTime);
@@ -44,6 +44,7 @@ public:
     virtual int ChooseTimeLimit(){return 0;}
 
 private:
+    class TimeBombApp* mApp;
     int mHitPoint;
     int mPendingBombNum;
     class Field* mCurrentField;
